Add special handling to remove broken connections
We allow only one connection from an ISM8. If this connection
breaks we don't receive any new packets, but a new a new connection
is created. We will continue to try to read data from the old
connection for some time and drop it when no data comes in.
Afterwards the new connection is accepted and the communication
works again.

The easiest way to test this, is to unplug the ISM8 and plug it
in again. In a similar way a dropped connection because of network
problems should be detected.