Introduce a online state which gets send over UDP and MQTT
The online state is send when the first data from the ISM is sent
and reset when the connection is closed (this can take some time
to be detected).
In addition the a MQTT LWT is set to also reset the online state
when the plugin crashes.

In addition to this basic online checking, there is now also
a new online_timeout setting, which indicates how long the plugin
should wait for new data from the ISM until it marks it as offline.